Create `DatabaseService` class with only `init` method
`DatabaseService` class represents the database that stores
accounts, agents and API keys data in this RAG application.

`init` method is used to run any prerequisites queries in the
like creating the table and adding extentions if needed.
For now, the method runs one query to create "uuid-ossp"
extention if is not exists, which is used to generate UUIDs.