Add two methods to `DatabaseService` class
Add `query` which is a getter method returns `queryObject`
method of `PostgresClient` instance to provide the ability
to perforem queries on the postgres database through
`DatabaseService` class.

Add `createTransaction` method which uses
`createTransaction` method of `PostgresClient` instance
to create and return a transaction on the postgres database
through `DatabaseService` class.