Create `VectorDatabaseService` service class
Create `VectorDatabaseService` service class which
represents the vector database that stores instructions
as vectors (so that it can searches for specific kind of
instructions efficiently) in RAG systems.

For now, `VectorDatabaseService` class has two methods
and two dependencies and one private property.

The two dependencies are an inctance of Qdrant client
and an inctance of `EmbeddingService` class.

The Methods are:

- `createDatasetsCollection` method to create the collection
of the instructions and create an index on "datasetId" field.

- `init` method to call inside it any required process before
`VectorDatabaseService` class becomes ready to use (calls
only `createDatasetsCollection` method for now).

The only property is `datasetsCollection` which holds
the name of datasets collection.