Create `ObjectStorageService` class
Create `ObjectStorageService` class with two methods,
one dependency and one property

The only dependency is an instance of `MinioClient` class from
`minio` client package to deal with `MimIO` object storage
database.

The only property is `buckets`, which is an object contains
the names of the used buckets in the application to store
the different data types, the purpose of this property is to
provide a unified way to select the needed backet to store
data in and to avoid writing buckets names hardcoded
in the other services.

The two methods are:

- `init` method to run any required task before the service
becomes ready to use.

- `createBuckets` method takes the buckets names from
`buckets` property and creates them in the object storage
database before service is used (called inside `init` method).