Create `AgentsService` class with two dependencies
Create `AgentsService` class with instances of `DatabaseService`
and `ObjectStorageService` classes as dependencies.

And also create `agentRowFieldsNamesMap` object which
is a map used to convert the camel case syntax that is used
in the code base to snake case when dealing with the database
because 'agents` table schema fields are written in snake case.