Create auth service with a basic 'log in root user' feature
Create `AuthService` class with a very basic and simple
'log in root user' feature to provide a straightforward way
to enter the system as root user.

`AuthService` class depends only on `JwtService` service
to generate the Json Web Token after successful log in,
And has two methods,

- `logInAsRoot` method to check the entered root username
and password and either to generate and return the JWT or
to return 'false'

- `getRootData` method which return the data of the root
user.