Create `chatResponseHandler` function
`chatResponseHandler` function is responsible for
handling the responses returned from `ollama.chat`
method (LLMs answer).

The function transforms the response stream into a
new readable stream (that supposed to be served  to
the client which sent the chat request via The API),

While transforming the response stream into the
new readable stream, The function collects the
response text to be saved in chat's record in the
database.