Create `chat` method in `ChatsService` class
`chat` method is the core part in chatting with agent
process in this RAG system.

It handles calling `fetchInstructions` method to get the
similar instructions to user's prompt,
Uses `systemMessageTemplete` to create the custom
system message of the agent, Calls `chat`  method of
LLM Service and passes to it chat messages with the
system message, And finally takes the response from
`chat`  method of LLM Service and uses
`chatResponseHandler` function to handle or process
the response.