Create `create` method in `DatasetsService` class
`create` method which is responsible to create a dataset
for an agent.

The function creates dataset row in datasets table, sets the
`dataset_id` field of the agent to the id of the created dataset yet,
Uploads the dataset file onto datasets bucket on the object storage,
And lastly posts a message to dataset processing worker to process
the dataset.