Rewrite `AuthService` class
Remove `logInAsRoot` and `getRootUserData` methods
from `AuthService` class and add `signUpUser` and
`signInUser` methods.

Why?, Because the system was designed to serve a single user
and this user enters the system via `getRootUserData` and
`logInAsRoot` methods, But now the idea was changed and this
system will be changed to serve many users and each user has
to sign up to use system through the web app.