Create API routes bootstrapping function
This function takes the required dependencies (which
are all controller + `guardService` object) to build the API
routes and injects into each route builder what it is depend
on, And then takes the result of each route builder function
(which is an instance of `Hono` class handles a specific API
route) and assign them to the API routes that they should
handle.

In the end the function returns an instance of `Hono`
class representing the API routes of the server.