Update `ApiKeysService` class to respect users plans limitations
Update `ApiKeysService.create` method to check if the user has reached
the limit of his plan in creating API Keys or not; and to increment the
number of user's API Keys after creating a new one.

Update `ApiKeysService.delete` method to decrement the number of
user's API Keys after deleting one.