Remove `inferences_count` column from users table
Tracking how many inferences the users have made every day
with a relational database (in our case postgres) will not be as
efficient as it would be when using a KV database.
So, Remove this column from from users table in the relational
database in favor of Deno KV database.