Update how the vector database service processes instructions
- Make `insert` method creates a new uuid for each instruction instead
of receiving a mongodb object id and transfering it into uuid.
And also make this method adds `createdAt` and `updatedAt` to
the payload of the instruction point inside the vector database.

- Make `update` method uses the incomming ids with the instructions
instead of transfering them into uuids because they should be valid
uuids in the first place because `insert` method now creates a uuid
for each instruction. Also make this method update the `updatedAt`
field of all updated instructions.

- Make `remove` method uses the array of instructions ids directly
without trying to transfer them into uuids because they should be
valid uuids in te first place.

- Use `wait: true` qdrant option with all insert, update and delete
operations to not return any response to the users until the
operations really complete.

- Delete `objectIdToUUID` helper function with its tests because it
is not needed anymore to transfer or convert mongodb object ids
to uuids.