Create `GET /api/instructions` endpoint to query instructions of a dataset
First create `queryInstructionsParamsValidator` middleware to parse and
validate the passed query parameters.

Then create `list` and `search` methods in `InstructionsService` service class
to provide the ability to eather list/explore instructions or search for instructions
semanticly.

Then create `query` method in `InstructionsController` controller class, and
assign it to its route with "read" permission in `instructionsRoutes.ts` file
This method should handle the requests to this endpoint and routes to eather
`instructionsService.search` service or `instructionsService.list` service based
on `searchText` query param whether it's passed or not.