Create an HTML template for sending OTP mails to users
After creating an HTML template for sending OTP mails to users,
Create a script that bundles this HTML template into a minified
version can be sent faster.

Create `src/constant/html-mails-templates/` directory and put
inside it `.gitkeep` file to be able to push thedirectory to the
remote repository so the deploy workflow can find the directory
and bundle the HTML templates into it.