Add email verification steps to the auth service
- Create `generateOTP` helper function that generates the
random OTPs.

- Create `createOneTimePasswordMail` helper function which
takes an OTP and injects it into the OTP HTML mail template and
returns a ready HTML mail to be sent to some user.

- Add `generateAndSendOTP` method to `AuthService` class.
This method does the following:
    - Uses `generateOTP` function to generate an OTP
    - Uses `createOneTimePasswordMail` function to create an OTP HTML mail
      from an existing template
    - Uses `randomString` function to generate a signature used to verify the
      sent OTP for verification.
    - Saves the generate OTP and signature associated with user's email in OTPs store
    - And finally uses an instance of `MailsSenderService` class to send the OTP HTML
      mail to the specified user's email address.

- Add `verifyOTP` method to `AuthService` class. This method is used to verify
the sent back OTP with its signature by the user, If they match the stored OTP
with its signature, The email of the user will be added to `verifiedEmails` set for
a period of time so the user can sign up within that time.