Enhance cold-start time by removing some unnecessary processes
Remove the `init` method of the database service that checks all tables
in the database every time the server starts, And that is completely
unnecessary, slows it down server and can not even apply new changes
on the schemas in the database if there are changes.

Instead, Use a database migration tool (dbmate) that runs only once in
CI/CD workflow to sync schemas of the tables in the database with the
latest changes if there are changes.