Enhance start up time by removing some unnecessary checks
Move the init method that checks if all necessary S3 buckets are existing
from object storage service to a separate script called `initS3Bukets.ts`.

Why?, Because calling the init method to check buckets that are already
existing every time the server start  hurts the start up time of the server
and this leads to bad user experience.