Removing some checks that hurts the start time of the server
Move the init method of the vector database service (that checks if
"datasets" collection in the vector DB exists and checks if the right
indecies exist) to a separate script file called `initVectorDatabase.ts`.

The goal is to not let these unnecessary checks run every time the
server start because these checks hurt the cold-start of the server in
case of hosting the server in a serverless solution (as we do on Deno
Deploy platform).