Log background tasks performence (how long each task take)
First move `performanceInSeconds` function to a separate file
in `helper` directory to be imported and reused anywhere in the
code base as helper function.

Then create `taskPerformenceLogger` method in the background
tasks service, This method is uses `performanceInSeconds` helper
function to calculate and log how long a giving task takes to complete.