Provide 2 options for clearing instructions from the vector DB
Make the `clear` method of the vector database service dynamically
clears eather the instructions of a single dataset (when both user's
email and dataset id are passed) or all instructions under the giving user
account (when only user's email was passed).

Then create two methods which are `clearDatasetInstructions` and
`clearUserInstructions` that use the `clear` method to clear the
targeted range of instructions.