Stop proxying uploading avatars to the storage in the server
Stop the idea of uploading avatar files to the storage through the server
and let the upload of users and agents avatars happen from the client side
to the cloud storage, and only send the server the urls of the uploaded
avatar store them.

This could save lots of bandwidth for the server.